the compiler caught me relying on implicit conversion from doubles to long, so cast explicitly instead
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@7416 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/lib/ztime_test.cc
@@ -77,10 +77,10 @@ main (int, char *argv[])
     errors += check (cond.c_str(), (one == two), __LINE__);
 
     // Test normalization.  These should all be equal.
-    ZTime norm (seconds+2, 500000);
+    ZTime norm ((long)seconds+2, 500000);
     ZTime t1 = ZTime((long)seconds, 2500000);
     TEST(t1 == norm);
-    ZTime t2 = ZTime(seconds+3, -500000);
+    ZTime t2 = ZTime((long)seconds+3, -500000);
     TEST(t2 == norm);
 
     // Try our hand at decoding.
